backend/requirements.txt
fastapi[standard]>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
chromadb>=0.5.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
pypdf>=4.0.0
websockets>=12.0
